+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add `CreateOptions::create()` as shorthand for `MemFile::create()`.
 * Change `CreateOptions::huge_tlb()` to accept any `Into<Option<HugeTlb>>`.
 * Take and return `CreateOptions` by value.
+* Add functions to create a `MemFile` that don't allocate a `CString` for the name.
 
 # v0.1.2 - 2021-02-02
 * Added `MemFile::into_file()` for safe ineroperability with crates that expect an `std::fs::File`.